move safe_state to state/mod
